Conditional intercept (#25)
Whilst working on a [Shields issue](https://github.com/badges/shields/issues/2469), I figured it would be nice to have a function allowing to flexibly opt in and out of an intercept based on a condition.

For instance, if an API token is available at runtime, forward to the real endpoint, otherwise intercept the request and mock it. The new `conditionalIntercept` addresses this.